refactor: update module exports, update imports
