+++ src/Data/index.js
@@ -1,5 +1,11 @@
-export { default as useBubblesData } from "./useBubblesData";
-export { default as useChoroplethData } from "./useChoroplethData";
-export { default as useDataExtents } from "./useDataExtents";
-export { default as useDataStatus } from "./useDataStatus";
-export { default as DataProvider } from "./DataProvider";
+export { default as useBubblesData } from "./hooks/useBubblesData";
+export { default as useChoroplethData } from "./hooks/useChoroplethData";
+export { default as useDataExtents } from "./hooks/useDataExtents";
+export { default as useDataFlags } from "./hooks/useDataFlags";
+export { default as useSummaryData } from "./hooks/useSummaryData";
+export { default as useLocationsData } from "./hooks/useLocationsData";
+export { default as useRegionGeojson } from "./hooks/useRegionGeojson";
+export { default as DataProvider } from "./components/DataProvider";
+export { default as DownloadDataButton } from "./components/DownloadDataButton";
+export { default as DataFlag } from "./components/DataFlag";
+export { default as DataFlags } from "./components/DataFlags";
