--- src/Legend/index.js
@@ -1,2 +1,0 @@
-export { default as BubbleLegend } from "./components/BubbleLegend";
-export { default as ChoroplethLegend } from "./components/ChoroplethLegend";
