+++ src/Locations/hooks/useLocationLoader.js
@@ -1,5 +1,5 @@
 import { useLocationStore } from "..";
-import { useRegionsGeojson } from "../../Data/useRegionGeojson";
+import { useRegionGeojson } from "../../Data";
 import { getRegionFromId } from "../utils";
 
 export default function useLocationLoader() {
@@ -17,7 +17,7 @@ export default function useLocationLoader() {
       if (all.indexOf(current) === -1) all.push(current);
       return all;
     }, []);
-  const regionsGeojson = useRegionsGeojson(regionIds, "choropleth");
+  const regionsGeojson = useRegionGeojson(regionIds, "choropleth");
   if (loadQueue.length === 0) return;
   const matches = [];
   const loadQueueIds = loadQueue.map((l) => l.id);
