+++ src/Map/index.js
@@ -1,1 +1,3 @@
 export { default as Map } from "./Map";
+export { default as BubbleLegend } from "./components/BubbleLegend";
+export { default as ChoroplethLegend } from "./components/ChoroplethLegend";
